refactor: :recycle: rename domain service directory and update method signatures for clarity
